New function

+++ lash/ExtraTools/bonus.py
@@ -10,12 +10,15 @@ config = config()
 @click.option('-s', type=click.INT, default=0, help='Randomization start number')
 @click.option('-e', type=click.INT, default=9, help='Randomization end number')
 @click.option('-c', type=click.INT, default=5, help='Number of characters in random output')
-def random(s=0, e=9, c=5):
+@click.option('-nl', is_flag=True, help='show to numbers left to randomize')
+def random(s, e, c, nl):
     if s != 0:
         e = s + 9
     randstr = ''
     for r in range(0, c):
         randstr += str((randint(s, e)))
+        if nl:
+            print(f'Numbers left: {c - (r + 1)}', end='\r')
     if len(randstr) > c:
         click.echo(randstr[0:c + 1])
     else:
@@ -39,6 +42,9 @@ def taskkiller():
     - None process find in <black_list>
     To use this command correctly, you need add processes 
     name manually in [/config.py]. To more information's use --help""")
-    for p in config['black_list']:
-        system(f'taskkill /F /IM {p} /T') if name == 'nt' else system(f'pkill {p}')
+    if name == 'nt':
+        for p in config['black_list']:
+            system(f'taskkill /F /IM {p} /T') if name == 'nt' else system(f'pkill {p}')
+    else:
+        ...
     print('\n<FUNCTION END>')
